Implement the review session finish modal
