+++ frontend/less/study.less
@@ -77,6 +77,21 @@
 	margin-top: 45px;
 }
 
-.navi-face {
-	width: 300px;
+.ui.modal > .content {
+	.image {
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+
+		.navi-face {
+			width: 300px;
+			height: 300px;
+		}
+
+		.navi-face-description {
+			font-weight: 400;
+			font-size: 14px;
+			margin-top: 0;
+		}
+	}
 }
