+++ frontend/src/tsconfig.json
@@ -9,6 +9,7 @@
 		"index.ts",
 		"all-words.ts",
 		"study.ts",
+		"study-lib.ts",
 		"review.ts"
 	],
 	"include": [
