+++ frontend/templates/study-session.ejs
@@ -27,17 +27,15 @@
 			<div class="header">
 				Congratulations, you finished Lesson <%= lesson['id'] + 1 %>!
 			</div>
-			<div class="image content">
-				<img class="image navi-face" src="/images/study/nitram.png" alt="A happy-looking Na'vi face">
+			<div class="content">
 				The words have been added to your review queue. You can start your reviews right away, or you can first view an overview of the words you learned.
 			</div>
 			<div class="actions">
-				<a class="ui primary button" id="to-review-button" href="/study/review?c=<%= course['id'] + 1%>&l=<%= lesson['id'] + 1 %>">Review &rarr;</a>
-				<a class="ui button" id="to-overview-button" href="/study/lesson?c=<%= course['id'] + 1%>&l=<%= lesson['id'] + 1 %>">Overview &rarr;</a>
+				<a class="ui primary button" id="to-review-button" href="/study/review?c=<%= course['id'] + 1%>&l=<%= lesson['id'] + 1 %>">Review</a>
+				<a class="ui button" id="to-overview-button" href="/study/lesson?c=<%= course['id'] + 1%>&l=<%= lesson['id'] + 1 %>">Overview</a>
 			</div>
 		</div>
 		<div id="word-info-modal" class="ui modal">
-			<div class="header"></div>
 			<div class="content"></div>
 			<div class="actions">
 				<a class="ui primary deny button">&rarr; <%- _('dialog-close') %></a>
